trajectory overlay
