+++ packages/react-components/lib/map/default-robot-marker.tsx
