+++ packages/react-components/lib/map/trajectory-marker.tsx
@@ -3,18 +3,13 @@ import React from 'react';
 import { Trajectory, trajectoryPath } from './trajectory';
 import { FollowAnimationPath } from './trajectory-paths';
 
-const debug = Debug('Robots:TrajectoryMarker');
-
-export type Conflict = number[];
+const debug = Debug('Map:TrajectoryMarker');
 
 export interface TrajectoryMarkerProps {
   trajectory: Trajectory;
-  conflict: boolean;
   color: string;
-  /**
-   * default: false
-   */
-  animationLoop?: boolean;
+  conflict?: boolean;
+  loopAnimation?: boolean;
   /**
    * default: 1
    */
@@ -22,15 +17,17 @@ export interface TrajectoryMarkerProps {
 }
 
 export const TrajectoryMarker = React.forwardRef(
-  (props: TrajectoryMarkerProps, ref: React.Ref<SVGGElement>) => {
-    const {
+  (
+    {
       trajectory,
-      conflict,
       color,
-      animationLoop = false,
+      conflict = false,
+      loopAnimation = false,
       animationScale = 1,
       ...otherProps
-    } = props;
+    }: TrajectoryMarkerProps,
+    ref: React.Ref<SVGGElement>,
+  ) => {
     debug(`render ${trajectory.id}`);
     const footprint = trajectory.dimensions;
 
@@ -46,7 +43,7 @@ export const TrajectoryMarker = React.forwardRef(
           color={color}
           footprint={footprint}
           conflict={conflict}
-          animationLoop={animationLoop}
+          loopAnimation={loopAnimation}
           animationScale={animationScale}
         />
       </g>
