+++ packages/react-components/lib/robots/index.ts
@@ -4,6 +4,4 @@ export * from './robot-accordion';
 export * from './robot-info';
 export * from './robot-panel';
 export * from './robot-table';
-export * from './trajectory';
-export * from './trajectory-marker';
 export * from './utils';
